Revert "4.22*: Blog List Expansion, step 10"
This reverts commit a1d263958d42b87e5c49dce21eeeb4e0f5f8dac1.

+++ part4/blog/controllers/blogs.js
@@ -2,16 +2,21 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
-const { userExtractor } = require('../utils/middleware')
+const tokenExtractor = require('../utils/middleware').tokenExtractor
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', {username: 1, name:1, id: 1})
   response.json(blogs)
 })
 
-blogsRouter.post('/', userExtractor, async (request, response) => {
+blogsRouter.post('/', tokenExtractor, async (request, response) => {
   const body = request.body
-  const user = request.user
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if(!decodedToken.id){
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
   if(!body.title || !body.url){
     return response.status(400).end()
@@ -31,9 +36,13 @@ blogsRouter.post('/', userExtractor, async (request, response) => {
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+blogsRouter.delete('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
-  const user = request.user
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!decodedToken.id){
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
   if(user.id.toString() === blog.user.toString()){
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
